// front/src/app/models/employ.ts
export interface Employ {
  _id: string,
  name: string;
  lastname: string;
  firstname: string;
  linkedin: string;
  snippet: string;
  company: string;
  image: string;
  email_perso: string;
  email_pro: string;
  phone: string;
  hobby: string;
  social_account: string;
  web_account: string;
  images: string;
  role: string;
  location: string;
}